requirements.txt
Flask>=3.0.0,<4.0.0
Flask-SQLAlchemy>=3.1.1,<4.0.0
SQLAlchemy>=2.0.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
google-auth>=2.29.0,<3.0.0
google-auth-oauthlib>=1.2.0,<2.0.0
google-api-python-client>=2.126.0,<3.0.0
beautifulsoup4>=4.12.3,<5.0.0
